@ranocha 2024-03-11
Note that `@batch` will also replace arrays by `PtrArray`s:

```julia
julia> using Polyester

julia> function foo!(x)
           @batch for i in eachindex(x)
               if i == 1
                   @show typeof(x)
               end
               x[i] = sin(i)
           end
           return sum(x)
       end
foo! (generic function with 1 method)

julia> foo!(rand(5))
typeof(x) = StrideArraysCore.PtrArray{Float64, 1, (1,), Tuple{Int64}, Tuple{Nothing}, Tuple{Static.StaticInt{1}}}
```