@svchb 2024-03-11
The current conclusion of #453 is that its an ARM Julia problem.