@efaulhaber 2024-03-12
So what do we do now? Ignore #453 and this PR because it's only relevant to macOS AMR? I'll try to find a MWE to report it at StrideArrays.jl, but I don't have time for this now.