@svchb 2024-10-03
Some of the issues probably also were the fixed bugs with the unsafe bitsets.
So this seems to be no longer relevant?
@efaulhaber close?